backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1

PyMySQL==1.1.1
python-dotenv==1.0.1

marshmallow==3.21.1

# Tests
pytest==8.3.3
pytest-cov==5.0.0